move test and fuzzing logic
