+++ trie-db/fuzz/fuzz_targets/no_ext_insert_rem.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::fuzz_that_no_extension_insert_remove;
+use trie_db_test::fuzz::fuzz_that_no_extension_insert_remove;
 
 fuzz_target!(|data: &[u8]| {
 	// fuzzed code goes here
