+++ trie-db/fuzz/fuzz_targets/prefix_iter.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::fuzz_prefix_iter;
+use trie_db_test::fuzz::fuzz_prefix_iter;
 
 fuzz_target!(|data: &[u8]| {
 	fuzz_prefix_iter::<reference_trie::NoExtensionLayout>(data);
