+++ trie-db/fuzz/fuzz_targets/prefix_seek_iter.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::{fuzz_prefix_seek_iter, PrefixSeekTestInput};
+use trie_db_test::fuzz::{fuzz_prefix_seek_iter, PrefixSeekTestInput};
 
 fuzz_target!(|data: PrefixSeekTestInput| {
 	fuzz_prefix_seek_iter::<reference_trie::SubstrateV1<reference_trie::RefHasher>>(data);
