+++ trie-db/fuzz/fuzz_targets/trie_root.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::fuzz_that_reference_trie_root;
+use trie_db_test::fuzz::fuzz_that_reference_trie_root;
 
 fuzz_target!(|data: &[u8]| {
 	fuzz_that_reference_trie_root::<reference_trie::NoExtensionLayout>(data);
