+++ trie-db/fuzz/fuzz_targets/trie_root_fix_len.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::fuzz_that_reference_trie_root_fix_length;
+use trie_db_test::fuzz::fuzz_that_reference_trie_root_fix_length;
 
 fuzz_target!(|data: &[u8]| {
 	fuzz_that_reference_trie_root_fix_length::<reference_trie::NoExtensionLayout>(data);
