+++ trie-db/fuzz/fuzz_targets/trie_root_new.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use trie_db_fuzz::fuzz_that_compare_implementations;
+use trie_db_test::fuzz::fuzz_that_compare_implementations;
 
 fuzz_target!(|data: &[u8]| {
 	// fuzzed code goes here
